+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Formal\ORM\Definition\Type\PointInTimeType::new()`
+
+### Deprecated
+
+- `Formal\ORM\Definition\Type\PointInTimeType::of()` as it uses a non standard string format. Use `::new()` instead, but don't forget to migrate your data.
+
+### Fixed
+
+- Psalm was complaining of a missing argument when using `PointInTimeType::of()`
+
 ## 3.1.0 - 2024-07-26
 
 ### Added